requirements.txt
azure-cognitiveservices-speech>=1.48.2
openai>=2.26.0
pandas>=2.3.3
python-dotenv>=1.2.2
streamlit>=1.55.0
pytest>=7.4.0
pytest-mock>=3.12.0
